+++ lib/concierge/suppliers/kigo/host_check.rb
@@ -7,15 +7,26 @@ module Kigo
   class HostCheck
     include Concierge::JSON
 
-    attr_reader :property_id, :request_handler
+    attr_reader :property_ids, :request_handler
 
-    def initialize(property_id, request_handler)
-      @property_id     = property_id
+    def initialize(property_ids, request_handler)
+      @property_ids    = property_ids
       @request_handler = request_handler
     end
 
     def active?
-      result = http.post(endpoint, json_encode(fake_params.value), { "Content-Type" => "application/json" })
+      res = nil
+      active = property_ids.any? do |id|
+        res = property_active?(id)
+        res.success? && res.value == true
+      end
+      return Result.new(true) if active
+      return Result.new(false) if res.success?
+      res
+    end
+
+    def property_active?(property_id)
+      result = http.post(endpoint, json_encode(fake_params(property_id).value), { "Content-Type" => "application/json" })
 
       if result.success?
         payload = json_decode(result.value.body)
@@ -36,7 +47,7 @@ module Kigo
       request_handler.http_client
     end
 
-    def fake_params
+    def fake_params(property_id)
       params = {
         property_id: property_id,
         check_in:    date.to_s,