+++ lib/concierge/suppliers/waytostay/quote.rb
@@ -8,7 +8,8 @@ module Waytostay
     UNAVAILBLE_ERROR_MESSAGES = [
       "Apartment is not available for the selected dates",
       "The minimum number of nights to book this apartment is",
-      "Cut off days restriction"
+      "Cut off days restriction",
+      "Max days for arrival restriction"
     ].freeze
     REQUIRED_RESPONSE_KEYS = [
       "booking_details.property_reference",
@@ -47,7 +48,7 @@ module Waytostay
           Result.new(Quotation.new(quote_params_from(response)))
         else
           augment_missing_fields(missing_keys)
-          Result.error(:unrecognised_response)
+          Result.error(:unrecognised_response, "Missing keys: #{missing_keys}")
         end
 
       elsif unavailable?(result) # for waytostay, unavailable is returned as a 422 error