+++ spec/lib/concierge/suppliers/ciirus/commands/cancel_spec.rb
@@ -27,6 +27,7 @@ RSpec.describe Ciirus::Commands::Cancel do
 
         expect(result).not_to be_success
         expect(result.error.code).to eq :savon_error
+        expect(result.error.data).to be_nil
       end
     end
 
@@ -50,6 +51,9 @@ RSpec.describe Ciirus::Commands::Cancel do
 
         expect(result.success?).to be false
         expect(result.error.code).to eq(:soap_error)
+        expect(result.error.data).to eq(
+          "(soap:Server) Server was unable to process request. ---> You do not have access to this booking"
+        )
       end
     end
   end