+++ spec/lib/concierge/suppliers/ciirus/commands/descriptions_html_fetcher_spec.rb
@@ -25,6 +25,7 @@ RSpec.describe Ciirus::Commands::DescriptionsHtmlFetcher do
 
         expect(result).not_to be_success
         expect(result.error.code).to eq :savon_error
+        expect(result.error.data).to be_nil
       end
     end