+++ spec/lib/concierge/suppliers/ciirus/commands/properties_fetcher_spec.rb
@@ -26,6 +26,7 @@ RSpec.describe Ciirus::Commands::PropertiesFetcher do
 
         expect(result).not_to be_success
         expect(result.error.code).to eq :savon_error
+        expect(result.error.data).to be_nil
       end
     end