+++ spec/lib/concierge/suppliers/ciirus/commands/property_rates_fetcher_spec.rb
@@ -54,6 +54,7 @@ RSpec.describe Ciirus::Commands::PropertyRatesFetcher do
 
         expect(result).not_to be_success
         expect(result.error.code).to eq :savon_error
+        expect(result.error.data).to be_nil
       end
     end