+++ spec/lib/concierge/suppliers/kigo/cancel_spec.rb
@@ -20,6 +20,7 @@ RSpec.describe Kigo::Cancel do
 
       expect(result).not_to be_success
       expect(result.error.code).to eq :connection_timeout
+      expect(result.error.data).to be_nil
     end